refactor: InputView 로직 변경
- InputView를 private 생성자를 이용해 인스턴스화 방지
- InputView에서 Scanner 클래스변수를 이용해 기능별로 입력받는 static 메서드 생성
