+++ src/main/java/racingcar/RacingCarMain.java
@@ -1,12 +1,17 @@
 package racingcar;
 
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
+
 import java.util.ArrayList;
 
 public class RacingCarMain {
     public static void main(String[] args) {
-        InputView inputView = new InputView();
+
+        int number = InputView.inputNumber();
+        int round = InputView.inputRound();
         ResultView resultView = new ResultView();
-        RacingCar racingCar = new RacingCar(inputView.number(), inputView.round());
+        RacingCar racingCar = new RacingCar(number, round);
 
         ArrayList<ArrayList<Integer>> result = racingCar.play();
